Relocate server tests
