+++ test/server/gameManagerSpec.js
@@ -1,5 +1,5 @@
 var expect = require("chai").expect;
-var gameManager = require("../server/gameManager.js");
+var gameManager = require("../../server/gameManager.js");
 
 describe("GameManager", function () {
